Implementado sistema de tag -Commit anteriro a la creación del frontend

+++ backend/src/controllers/actividadTag.controller.ts
@@ -1,10 +1,12 @@
 import type {Request, Response} from 'express';
 import * as ActividadTagModel from '../models/actividadTag.model.js';
 import {eliminarActividadTag} from "../models/actividadTag.model.js";
+import * as TagService from "../services/tag.service.js";
+import * as ActividadService from "../services/actividad.service.js";
 
 
 
-export const getTodosActividadTags = async (res: Response) => {
+export const getTodosActividadTags = async (req: Request ,res: Response) => {
     try {
         const actividadTags = await ActividadTagModel.getAllActividadTags();
         res.json(actividadTags);
@@ -25,9 +27,35 @@ export const getTagPorActividad = async (req: Request, res: Response) => {
 };
 export const createActividadTag = async (req: Request, res: Response) => {
 
-    //Existe el tag
+    const idUsuario = req.userId;
+    const idActividad = req.body.idActividad;
+    const existeTag = await TagService.TagService.existeTagPorid(req.body.idTag);
+    const existeActividad = await ActividadService.ActividadService.existeActividad(idActividad);
 
+    //valida los id usuario
+    if (!idUsuario || isNaN(idUsuario)) {
+        return res.status(400).json({ message: 'ID de usuario inválido' });
+    }
+    //Existe el tag
+    if (!existeTag) {
+        return res.status(400).json({ message: 'El tag no existe' });
+    }
     //existe la actividad
+    if (!existeActividad) {
+        return res.status(400).json({ message: 'La actividad no existe' });
+    }
+    //el usuario es el creador de la actividad
+    const esCreador = await ActividadService.ActividadService.esCreadorActividad(idActividad, idUsuario);
+    if (!esCreador) {
+        return res.status(403).json({ message: 'No tienes permiso para asignar tags a esta actividad' });
+    }
+    //existe la actividad tag
+    const actividadTags = await ActividadTagModel.getTagsActividad(idActividad);
+    const existeActividadTag = actividadTags.find(at => at.idTag === parseInt(req.body.idTag));
+    if (existeActividadTag) {
+        return res.status(400).json({ message: 'El tag ya está asignado a la actividad' });
+    }
+
 
     try {
         const actividadTag = await ActividadTagModel.crearActividadTag(req.body);
@@ -41,13 +69,35 @@ export const createActividadTag = async (req: Request, res: Response) => {
 //Eliminar un tag de una actividad por id
 export const deleteActividadTag = async (req: Request, res: Response) => {
 
+    const idUsuario = req.userId;
+    const idActividad = req.body.idActividad;
+    const existeTag = await TagService.TagService.existeTagPorid(req.body.idTag);
+    const existeActividad = await ActividadService.ActividadService.existeActividad(idActividad);
+
+    //valida los id usuario
+    if (!idUsuario || isNaN(idUsuario)) {
+        return res.status(400).json({ message: 'ID de usuario inválido' });
+    }
     //Existe el tag
-
+    if (!existeTag) {
+        return res.status(400).json({ message: 'El tag no existe' });
+    }
     //existe la actividad
+    if (!existeActividad) {
+        return res.status(400).json({ message: 'La actividad no existe' });
+    }
+    //el usuario es el creador de la actividad
+    const esCreador = await ActividadService.ActividadService.esCreadorActividad(idActividad, idUsuario);
+    if (!esCreador) {
+        return res.status(403).json({ message: 'No tienes permiso para asignar tags a esta actividad' });
+    }
 
-    //existe el tag actividad
-
-
+    //existe la actividad tag
+    const actividadTags = await ActividadTagModel.getTagsActividad(idActividad);
+    const existeActividadTag = actividadTags.find(at => at.idTag === parseInt(req.body.idTag));
+    if (!existeActividadTag) {
+        return res.status(400).json({ message: 'El tag no está asignado a la actividad' });
+    }
     try {
         await ActividadTagModel.eliminarActividadTag(parseInt(req.body.idActividad), parseInt(req.body.idTag));
         res.status(204).send();
